save/load time data to/from .sgf files (#29)
change details
- SaveSgfCommand and LoadGameCommand now write/read the SGF properties
  TM / OT (game info node) and BL / WL / OB / OW (move node)
- GoNodeTimeData no longer requires the isTimeDataForBlackPlayer value
  for initialization, instead the property is now writable; reason:
  handling in LoadGameCommand becomes much easier this way
- change data type of several properties to "unsigned long" to match the
  type SGFCNumber
  - GoNodeTimeData properties remainingNumberOfMoves and
    remainingNumberOfPeriods
  - GoPlayerTimeData properties remainingNumberOfMoves and
    remainingNumberOfPeriods
  - GoTimeSystem properties minimumNumberOfMovesPerPeriod and
    numberOfPeriods
  - TimeView properties remainingNumberOfMovesOrPeriods and
    remainingTimeInSecondsRoundedUp (internal)
- bug fix in GoTimeDataValidator when determining presence of absolute
  time system

+++ src/go/GoPlayerTimeData.h
@@ -50,7 +50,7 @@
 
 - (void) updateAfterMoveWasPlayed:(GoNodeTimeData*)nodeTimeData;
 - (void) updateAfterNodeChanged:(GoNode*)node;
-- (int) remainingNumberOfMovesOrPeriods;
+- (unsigned long) remainingNumberOfMovesOrPeriods;
 - (enum GoClockState) clockState;
 - (enum GoClockSuspendedReason) clockSuspendedReason;
 
@@ -74,12 +74,18 @@
 ///
 /// The default value after initialization depends on the parameters in the
 /// GoTimeSettings object that is supplied to the initializer.
-@property(nonatomic, assign, readonly) unsigned int remainingNumberOfMoves;
+///
+/// The property type must be able to hold an SGFCNumber, to avoid mismatches
+/// with the type of property @e remainingNumberOfMoves in GoNodeTimeData.
+@property(nonatomic, assign, readonly) unsigned long remainingNumberOfMoves;
 /// @brief The number of periods that the player has left after they played
 /// their most recent move.
 ///
 /// The default value after initialization depends on the parameters in the
 /// GoTimeSettings object that is supplied to the initializer.
-@property(nonatomic, assign, readonly) unsigned int remainingNumberOfPeriods;
+///
+/// The property type must be able to hold an SGFCNumber, to avoid mismatches
+/// with the type of property @e remainingNumberOfPeriods in GoNodeTimeData.
+@property(nonatomic, assign, readonly) unsigned long remainingNumberOfPeriods;
 
 @end
